refactor: remove duplicate notification validator
