+++ src/components/config/validation/validator/notification.ts
@@ -22,92 +22,79 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
+import { validator as dataStatuspageSchemaValidator } from '../../../../plugins/visualization/atlassian-status-page'
+import { validator as dataInstatusSchemaValidator } from '../../../../plugins/visualization/instatus'
+import { NotificationSendingError } from '../../../notification'
 import {
-  slug as atlassianStatuspageSlug,
-  validateConfig as atlassianStatuspageValidateConfig,
-} from '../../../../plugins/visualization/atlassian-status-page'
-import {
-  slug as instatusPageSlug,
-  validateConfig as instatusPageValidateConfig,
-} from '../../../../plugins/visualization/instatus'
-import { newPagerDuty, type Notification } from '../../../notification/channel'
-import { requiredFieldMessages } from '../notification-required-fields'
+  dataDingtalkSchemaValidator,
+  dataDiscordSchemaValidator,
+  dataGoogleChatSchemaValidator,
+  dataGotifySchemaValidator,
+  dataLarkSchemaValidator,
+  dataMailgunSchemaValidator,
+  dataMonikaNotifSchemaValidator,
+  dataOpsgenieSchemaValidator,
+  dataPushbulletSchemaValidator,
+  dataPushoverSchemaValidator,
+  dataSendgridSchemaValidator,
+  dataSlackSchemaValidator,
+  dataSMTPSchemaValidator,
+  dataTeamsSchemaValidator,
+  dataTelegramSchemaValidator,
+  dataWebhookSchemaValidator,
+  dataWorkplaceSchemaValidator,
+  newPagerDuty,
+  Notification,
+  validator,
+} from '../../../notification/channel'
 
-const checkRecipients = (notification: Notification): string | undefined => {
-  // check one-by-one instead of using indexOf or includes so the type is correct without type assertion
-  const notifData: any = notification.data
-  if (
-    (notifData?.recipients?.length ?? 0) === 0 &&
-    (notification.type === 'mailgun' ||
-      notification.type === 'smtp' ||
-      notification.type === 'sendgrid' ||
-      notification.type === 'whatsapp')
-  ) {
-    return 'Recipients does not exists or has length lower than 1!'
-  }
-}
-
-const validateRequiredFields = (
-  notification: Notification
-): string | undefined => {
-  const { data }: any = notification
-  const reqFields = requiredFieldMessages[notification.type]
-  if (!reqFields)
-    return `Notifications type is not allowed (${(notification as any)?.type})`
-
-  const keys = Object.keys(reqFields)
-  for (const field of keys) {
-    if (!data[field]) {
-      return reqFields[field]
-    }
-  }
-}
-
-const checkByNotificationType = (
-  notification: Notification
-): string | undefined => {
-  const pagerduty = newPagerDuty()
-  if (
-    notification.type === atlassianStatuspageSlug &&
-    atlassianStatuspageValidateConfig(notification.data)
-  ) {
-    return atlassianStatuspageValidateConfig(notification.data)
-  }
-
-  if (
-    notification.type === pagerduty.slug &&
-    pagerduty.validateConfig(notification.data)
-  ) {
-    return pagerduty.validateConfig(notification.data)
-  }
-
-  if (
-    notification.type === instatusPageSlug &&
-    instatusPageValidateConfig(notification.data)
-  ) {
-    return instatusPageValidateConfig(notification.data)
-  }
-
-  const missingField = validateRequiredFields(notification)
-  if (missingField) {
-    return missingField
-  }
-}
-
-export const validateNotification = (
+export const validateNotification = async (
   notifications: Notification[]
-): string | undefined => {
-  if (notifications.length === 0) return
-
-  for (const notification of notifications) {
-    const missingRecipient = checkRecipients(notification)
-    if (missingRecipient) {
-      return missingRecipient
-    }
-
-    const missingSlug = checkByNotificationType(notification)
-    if (missingSlug) {
-      return missingSlug
-    }
+): Promise<void> => {
+  const hasNotification = notifications.length > 0
+  if (!hasNotification) {
+    return
   }
+
+  const pagerduty = newPagerDuty()
+  const validators = {
+    desktop: validator,
+    dingtalk: dataDingtalkSchemaValidator,
+    discord: dataDiscordSchemaValidator,
+    'google-chat': dataGoogleChatSchemaValidator,
+    gotify: dataGotifySchemaValidator,
+    instatus: dataInstatusSchemaValidator,
+    lark: dataLarkSchemaValidator,
+    mailgun: dataMailgunSchemaValidator,
+    'monika-notif': dataMonikaNotifSchemaValidator,
+    opsgenie: dataOpsgenieSchemaValidator,
+    pagerduty: pagerduty.validator,
+    pushbullet: dataPushbulletSchemaValidator,
+    pushover: dataPushoverSchemaValidator,
+    sendgrid: dataSendgridSchemaValidator,
+    slack: dataSlackSchemaValidator,
+    smtp: dataSMTPSchemaValidator,
+    statuspage: dataStatuspageSchemaValidator,
+    teams: dataTeamsSchemaValidator,
+    telegram: dataTelegramSchemaValidator,
+    webhook: dataWebhookSchemaValidator,
+    whatsapp: dataWebhookSchemaValidator,
+    workplace: dataWorkplaceSchemaValidator,
+  }
+
+  await Promise.all(
+    notifications.map(async (notification) => {
+      const validator = validators[notification.type]
+
+      if (!validator) {
+        return
+      }
+
+      try {
+        await validator.validateAsync(notification.data)
+      } catch (error: any) {
+        throw NotificationSendingError.create(notification.type, error?.message)
+      }
+    })
+  )
 }
