--- src/utils/validate-response.ts
@@ -1,11 +1,0 @@
-import { Validation } from '../interfaces/validation'
-
-export const setInvalidResponse = (message: string): Validation => ({
-  valid: false,
-  message: message,
-})
-
-export const VALID_CONFIG: Validation = {
-  valid: true,
-  message: '',
-}
